Cleaning up

+++ src/mako/utility/UUID.php
@@ -142,7 +142,7 @@ class UUID
 	{
 		[$usec, $sec] = explode(' ', microtime());
 
-		$random = hex2bin(dechex($sec . substr($usec, 2, 5)) . bin2hex(random_bytes(10)));
+		$random = hex2bin(dechex((int) ($sec . substr($usec, 2, 5))) . bin2hex(random_bytes(10)));
 
 		$random[6] = chr(ord($random[6]) & 0x0f | 0x40);
